+++ internal/provider/resource_meraki_networks_syslog_servers.go
@@ -3,6 +3,7 @@ package provider
 // RESOURCE NORMAL
 import (
 	"context"
+	"strconv"
 
 	merakigosdk "github.com/meraki/dashboard-api-go/v3/sdk"
 
@@ -320,8 +321,9 @@ func ResponseNetworksGetNetworkSyslogServersItemToBodyRs(state NetworksSyslogSer
 					result[i] = ResponseNetworksGetNetworkSyslogServersServersRs{
 						Host: types.StringValue(servers.Host),
 						Port: func() types.Int64 {
-							if servers.Port != nil {
-								return types.Int64Value(int64(*servers.Port))
+							if servers.Port != "" {
+								serverPort, _ := strconv.Atoi(servers.Port)
+								return types.Int64Value(int64(serverPort))
 							}
 							return types.Int64{}
 						}(),